requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
pymysql>=1.1.0
anthropic>=0.28.0
openai>=1.30.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-multipart>=0.0.9
openpyxl>=3.1.2
pypdf>=4.0.0
